refactor: update the file name to redeems-fetcher
